backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
groq==0.15.0
pypdf==5.1.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.10.4
sse-starlette==2.2.1
rank-bm25==0.2.2
